+++ lib/utils/resolveExternalFilepath.js
@@ -23,6 +23,8 @@ module.exports = function resolveExternalFilepath(filepath, modulePath) {
       return `${srcPath}.${ext}`;
     }
   }
+
+  throw new Error(`Could not resolve one of filepaths: \n\t${extensions.map(ext => `${srcPath}.${ext}`).join('\n\t')}`);
 }
 
 // Thanks for the inspiration, @pasupuletics  https://github.com/pasupuletics/learning/blob/37d10ba5d2a9decb2d33e66441976e92910c7e26/react-docgen-external-proptypes-handler.js#L111
