+++ tests/fixtures/components/assets/api-1.js
