+++ tests/fixtures/components/assets/api-2.js
