+++ tests/fixtures/components/assets/api-3.js
