+++ tests/fixtures/components/assets/api-cjs-named-exports.js
