+++ tests/fixtures/components/assets/api-external-values.js
