+++ tests/fixtures/components/assets/external-default-string-propType.js
