+++ tests/fixtures/components/assets/external-propTypes.js
