+++ tests/fixtures/components/assets/external-values.js
