+++ tests/fixtures/components/component-local-spread.js
