+++ tests/fixtures/components/component-spreads.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { propTypes, defaultProps } from './api-1';
-import { propTypes as propTypesApi2, defaultProps as defaultPropsApi2 } from './api-2';
+import { propTypes, defaultProps } from './assets/api-1';
+import { propTypes as propTypesApi2, defaultProps as defaultPropsApi2 } from './assets/api-2';
 
 const ComponentSpreads = props => (<div>component-spreads</div>);
